Implemented abhay and jyotiram imple

+++ src/main/java/com/example/repository/CustomerQueryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.entity.CustomerQuery;
 
 @Repository
-public interface CustomerQueryRepository extends JpaRepository<CustomerQuery, Integer>{
+public interface CustomerQueryRepository extends JpaRepository<CustomerQuery, Long>{
 
 }
