+++ src/main/java/com/example/service/CompanyResponseFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.CompanyResponse;
 
 public interface CompanyResponseFetchService {
-	CompanyResponse fetchById(int id);
+	CompanyResponse fetchById(Long id);
 
 }
