+++ src/main/java/com/example/service/CustomerQueryUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.CustomerQuery;
 
 public interface CustomerQueryUpdateService {
-	CustomerQuery updateData(CustomerQuery customerQuery , int id);
+	CustomerQuery updateData(CustomerQuery customerQuery , Long id);
 
 }
