+++ src/main/java/com/example/service/FeedbackFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Feedback;
 
 public interface FeedbackFetchService {
-	Feedback fetchById(int id);
+	Feedback fetchById(Long id);
 
 }
