+++ src/main/java/com/example/service/ProductFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Product;
 
 public interface ProductFetchService {
-	Product fetchById(int id);
+	Product fetchById(Long id);
 
 }
