+++ src/main/java/com/example/service/ProductReviewFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.ProductReview;
 
 public interface ProductReviewFetchService {
-	ProductReview fetchById(int id);
+	ProductReview fetchById(Long id);
 
 }
