+++ src/main/java/com/example/service/SubCategoryDeleteService.java
@@ -1,6 +1,6 @@
 package com.example.service;
 
 public interface SubCategoryDeleteService {
-	void deleteById(int id);
+	void deleteById(Long id);
 
 }
