+++ src/main/java/com/example/service/TrackingFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Tracking;
 
 public interface TrackingFetchService {
-	Tracking fetchById(int id);
+	Tracking fetchById(Long id);
 
 }
