+++ src/main/java/com/example/service/TrackingUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Tracking;
 
 public interface TrackingUpdateService {
-	Tracking updateData(Tracking tracking , int id);
+	Tracking updateData(Tracking tracking , Long id);
 
 }
