+++ src/main/java/com/example/service/UpiUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Upi;
 
 public interface UpiUpdateService {
-	Upi updateData(Upi upi , int id);
+	Upi updateData(Upi upi , Long id);
 
 }
